% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_types.R
\name{absolute_error}
\alias{absolute_error}
\title{Calculate Absolute Error}
\usage{
absolute_error(actual, predicted, na.rm = FALSE)
}
\arguments{
\item{actual}{Numeric vector of actual (observed) values.}

\item{predicted}{Numeric vector of predicted values.}

\item{na.rm}{Logical. If \code{TRUE}, remove \code{NA} pairs before
computing. Default is \code{FALSE}.}
}
\value{
Numeric vector of absolute errors.
}
\description{
Compute absolute error between actual and predicted values,
as defined in Equation (2) of Agustini et al. (2026).
}
\examples{
actual <- c(10, 20, 30, 40, 50)
predicted <- c(11, 19, 32, 38, 51)
absolute_error(actual, predicted)

}
