% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thresholds.R
\name{print.al_threshold}
\alias{print.al_threshold}
\title{Print Method for al_threshold Objects}
\usage{
\method{print}{al_threshold}(x, ...)
}
\arguments{
\item{x}{An \code{al_threshold} object.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns the input object.
}
\description{
Print Method for al_threshold Objects
}
