% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_types.R
\name{squared_error}
\alias{squared_error}
\title{Calculate Squared Error}
\usage{
squared_error(actual, predicted, na.rm = FALSE)
}
\arguments{
\item{actual}{Numeric vector of actual (observed) values.}

\item{predicted}{Numeric vector of predicted values.}

\item{na.rm}{Logical. If \code{TRUE}, remove \code{NA} pairs before
computing. Default is \code{FALSE}.}
}
\value{
Numeric vector of squared errors.
}
\description{
Compute squared error between actual and predicted values,
as defined in Equation (1) of Agustini et al. (2026).
}
\examples{
actual <- c(10, 20, 30, 40, 50)
predicted <- c(11, 19, 32, 38, 51)
squared_error(actual, predicted)

}
\references{
Agustini, M., Fithriasari, K., & Prastyo, D.D. (2026). An accuracy-level
method for robust evaluation in predictive analytics. \emph{Decision
Analytics Journal}, 18, 100661. \doi{10.1016/j.dajour.2025.100661}
}
